+++ src/python/pants/engine/exp/scheduler.py
@@ -14,10 +14,15 @@ from pants.build_graph.address import Address
 from pants.engine.exp.addressable import StructAddress, parse_variants
 from pants.engine.exp.struct import Struct
 from pants.engine.exp.targets import Target, Variants
+from pants.util.meta import AbstractClass
 from pants.util.objects import datatype
 
 
-class Selector(object):
+class Selector(AbstractClass):
+  @abstractproperty
+  def optional(self):
+    """Return true if this Selector is optional. It may result in a `None` match."""
+
   @abstractmethod
   def construct_node(self, subject, variants):
     """Constructs a Node for this Selector and the given Subject/Variants.
@@ -26,8 +31,14 @@ class Selector(object):
     """
 
 
-class Select(datatype('Subject', ['product']), Selector):
-  """Selects the given Product for the Subject provided to the constructor."""
+class Select(datatype('Subject', ['product', 'optional']), Selector):
+  """Selects the given Product for the Subject provided to the constructor.
+
+  If optional=True and no matching product can be produced, will return None.
+  """
+
+  def __new__(self, product, optional=False):
+    return super(Select, self).__new__(self, product, optional)
 
   def construct_node(self, subject, variants):
     return SelectNode(subject, self.product, variants, None)
@@ -40,6 +51,7 @@ class SelectVariant(datatype('Variant', ['product', 'variant_key']), Selector):
   will only match when a consumer passes a variant value for "thrift" that matches the name of an
   ApacheThrift value.
   """
+  optional = False
 
   def construct_node(self, subject, variants):
     return SelectNode(subject, self.product, variants, self.variant_key)
@@ -51,6 +63,7 @@ class SelectDependencies(datatype('Dependencies', ['product', 'deps_product']), 
   The dependencies declared on `deps_product` will be provided to the requesting task
   in the order they were declared.
   """
+  optional = False
 
   def construct_node(self, subject, variants):
     return DependenciesNode(subject, self.product, variants, self.deps_product)
@@ -62,6 +75,7 @@ class SelectProjection(datatype('Projection', ['product', 'projected_product', '
   Projecting an input allows for deduplication in the graph, where multiple Subjects
   resolve to a single backing Subject instead.
   """
+  optional = False
 
   def construct_node(self, subject, variants):
     # Input product type doesn't match: not satisfiable.
@@ -78,6 +92,7 @@ class SelectProjection(datatype('Projection', ['product', 'projected_product', '
 
 class SelectLiteral(datatype('Literal', ['subject', 'product']), Selector):
   """Selects a literal Subject (other than the one applied to the selector)."""
+  optional = False
 
   def construct_node(self, subject, variants):
     # NB: Intentionally ignores subject parameter to provide a literal subject.
@@ -88,47 +103,45 @@ class SchedulingError(Exception):
   """Indicates inability to make a scheduling promise."""
 
 
-class NoProducersError(SchedulingError):
-  """Indicates no planners were able to promise a product for a given subject."""
-
-  def __init__(self, product_type, subject=None, configuration=None):
-    msg = ('No plans to generate {!r}{} could be made.'
-            .format(product_type.__name__,
-                    ' for {!r}'.format(subject) if subject else '',
-                    ' (with config {!r})' if configuration else ''))
-    super(NoProducersError, self).__init__(msg)
-
-
 class PartiallyConsumedInputsError(SchedulingError):
-  """No planner was able to produce a plan that consumed the given input products."""
+  """No task was able to consume a particular literal product for a subject, although some tried.
 
-  @staticmethod
-  def msg(output_product, subject, partially_consumed_products):
-    yield 'While attempting to produce {} for {}, some products could not be consumed:'.format(
-             output_product.__name__, subject)
-    for input_product, planners in partially_consumed_products.items():
-      yield '  To consume {}:'.format(input_product)
-      for planner, additional_inputs in planners.items():
-        inputs_str = ' OR '.join(str(i) for i in additional_inputs)
-        yield '    {} needed ({})'.format(type(planner).__name__, inputs_str)
+  In particular, this error allows for safe composition of configuration on targets (ie,
+  ThriftSources AND ThriftConfig), because if a task requires multiple inputs for a subject
+  but cannot find them, a useful error is raised.
 
-  def __init__(self, output_product, subject, partially_consumed_products):
-    msg = '\n'.join(self.msg(output_product, subject, partially_consumed_products))
-    super(PartiallyConsumedInputsError, self).__init__(msg)
+  TODO: Improve the error message in the presence of failures due to mismatched variants.
+  """
+
+  @classmethod
+  def _msg(cls, inverted_symbol_table, partially_consumed_inputs):
+    def name(product):
+      return inverted_symbol_table[product]
+    for subject, tasks_and_inputs in partially_consumed_inputs.items():
+      yield '\nSome products were partially specified for `{}`:'.format(subject)
+      for ((input_product, output_product), tasks) in tasks_and_inputs.items():
+        yield '  To consume `{}` and produce `{}`:'.format(name(input_product), name(output_product))
+        for task, additional_inputs in tasks:
+          inputs_str = ' AND '.join('`{}`'.format(name(i)) for i in additional_inputs)
+          yield '    {} also needed ({})'.format(task.__name__, inputs_str)
+
+  @classmethod
+  def create(cls, inverted_symbol_table, partially_consumed_inputs):
+    msg = '\n'.join(cls._msg(inverted_symbol_table, partially_consumed_inputs))
+    return cls(msg)
 
 
 class ConflictingProducersError(SchedulingError):
-  """Indicates more than one planner was able to promise a product for a given subject.
+  """Indicates that there was more than one source of a product for a given subject.
 
   TODO: This will need to be legal in order to support multiple Planners producing a
-  (mergeable) Classpath for one subject, for example.
+  (mergeable) Classpath for one subject, for example. see:
+    https://github.com/pantsbuild/pants/issues/2526
   """
 
-  def __init__(self, product_type, subject, plans):
-    msg = ('Collected the following plans for generating {!r} from {!r}:\n\t{}'
-            .format(product_type.__name__,
-                    subject,
-                    '\n\t'.join(str(p.func_or_task_type.value) for p in plans)))
+  def __init__(self, subject, product, matches):
+    msgs = '\n  '.join('{}: {}'.format(k, v) for k, v in matches.items())
+    msg = 'More than one source of {} for {}:\n  {}'.format(product.__name__, subject, msgs)
     super(ConflictingProducersError, self).__init__(msg)
 
 
@@ -138,12 +151,17 @@ class State(object):
     raise ValueError('Unrecognized Node State: {}'.format(state))
 
 
+class Noop(datatype('Noop', ['msg']), State):
+  """Indicates that a Node did not have the inputs which would be needed for it to execute."""
+  pass
+
+
 class Return(datatype('Return', ['value']), State):
   """Indicates that a Node successfully returned a value."""
   pass
 
 
-class Throw(datatype('Throw', ['msg']), State):
+class Throw(datatype('Throw', ['exc']), State):
   """Indicates that a Node should have been able to return a value, but failed."""
   pass
 
@@ -184,7 +202,7 @@ class Node(object):
 class SelectNode(datatype('SelectNode', ['subject', 'product', 'variants', 'variant_key']), Node):
   """A Node that selects a product for a subject.
 
-  A Select can be satisfied by multiple sources, and (TODO: currently) acts like an OR. The
+  A Select can be satisfied by multiple sources, but fails if multiple sources produce a value. The
   'variants' field represents variant configuration that is propagated to dependencies. When
   a task needs to consume a product as configured by the variants map, it uses the SelectVariant
   selector, which introduces the 'variant' value to restrict the names of values selected by a
@@ -209,8 +227,8 @@ class SelectNode(datatype('SelectNode', ['subject', 'product', 'variants', 'vari
         for configuration in candidate.configurations:
           yield configuration
 
-    # TODO: returning only the first configuration of a given type/variant. Need to define
-    # mergeability for products.
+    # TODO: returning only the first literal configuration of a given type/variant. Need to
+    # define mergeability for products.
     for item in items():
       if not isinstance(item, self.product):
         continue
@@ -218,19 +236,6 @@ class SelectNode(datatype('SelectNode', ['subject', 'product', 'variants', 'vari
         continue
       return item
     return None
-
-  def _task_sources(self, node_builder, variants):
-    """Returns a sequence of potential source Nodes for this Select."""
-    # Tasks.
-    for task_node in node_builder.task_nodes(self.subject, self.product, variants):
-      yield task_node
-    # An Address that can be resolved into a Struct.
-    # TODO: This node defines a special case for Addresses and Structs by recognizing that they
-    # might be-a Product after resolution, and so it begins by attempting to resolve a Struct for
-    # a subject Address. This type of cast/conversion should likely be reified.
-    if isinstance(self.subject, Address) and issubclass(self.product, Struct):
-      struct_address = StructAddress(self.subject.spec_path, self.subject.target_name)
-      yield SelectNode(struct_address, Struct, None, None)
 
   def step(self, dependency_states, node_builder):
     # Request default Variants for the subject, so that if there are any we can propagate
@@ -253,7 +258,7 @@ class SelectNode(datatype('SelectNode', ['subject', 'product', 'variants', 'vari
                         if key == self.variant_key] if variants else None
       if not variant_values:
         # Select cannot be satisfied: no variant configured for this key.
-        return Throw('Variant key {} was not configured in variants {}'.format(
+        return Noop('Variant key {} was not configured in variants {}'.format(
           self.variant_key, variants))
       variant_value = variant_values[0]
 
@@ -264,23 +269,33 @@ class SelectNode(datatype('SelectNode', ['subject', 'product', 'variants', 'vari
 
     # Else, attempt to use a configured task to compute the value.
     has_waiting_dep = False
-    dependencies = list(self._task_sources(node_builder, variants))
+    dependencies = list(node_builder.task_nodes(self.subject, self.product, variants))
+    matches = {}
     for dep in dependencies:
       dep_state = dependency_states.get(dep, None)
       if dep_state is None or type(dep_state) == Waiting:
         has_waiting_dep = True
         continue
       elif type(dep_state) == Throw:
+        return dep_state
+      elif type(dep_state) == Noop:
         continue
       elif type(dep_state) != Return:
         State.raise_unrecognized(dep_state)
       # We computed a value: see whether we can use it.
       literal_value = self._select_literal(dep_state.value, variant_value)
       if literal_value is not None:
-        return Return(literal_value)
+        matches[dep] = literal_value
     if has_waiting_dep:
       return Waiting(dependencies)
-    return Throw("No source of {}.".format(self))
+    elif len(matches) > 1:
+      # TODO: Multiple successful tasks are not currently supported. We should allow for this
+      # by adding support for "mergeable" products. see:
+      #   https://github.com/pantsbuild/pants/issues/2526
+      return Throw(ConflictingProducersError(self.subject, self.product, matches))
+    elif len(matches) == 1:
+      return Return(matches.values()[0])
+    return Noop('No source of {}.'.format(self))
 
 
 class DependenciesNode(datatype('DependenciesNode', ['subject', 'product', 'variants', 'dep_product']), Node):
@@ -311,7 +326,9 @@ class DependenciesNode(datatype('DependenciesNode', ['subject', 'product', 'vari
     if dep_product_state is None or type(dep_product_state) == Waiting:
       return Waiting([dep_product_node])
     elif type(dep_product_state) == Throw:
-      return Throw('Could not compute {} to determine dependencies.'.format(dep_product_node))
+      return dep_product_state
+    elif type(dep_product_state) == Noop:
+      return Noop('Could not compute {} to determine dependencies.'.format(dep_product_node))
     elif type(dep_product_state) != Return:
       State.raise_unrecognized(dep_product_state)
 
@@ -325,7 +342,9 @@ class DependenciesNode(datatype('DependenciesNode', ['subject', 'product', 'vari
         # of them.
         return Waiting([dep_product_node] + dependencies)
       elif type(dep_state) == Throw:
-        return Throw('Failed to compute dependency {}'.format(dependency))
+        return dep_state
+      elif type(dep_state) == Noop:
+        return Throw(ValueError('No source of explicit dependency {}'.format(dependency)))
       elif type(dep_state) != Return:
         raise State.raise_unrecognized(dep_state)
     # All dependencies are present! Set our value to a list of the resulting values.
@@ -341,21 +360,29 @@ class TaskNode(datatype('TaskNode', ['subject', 'product', 'variants', 'func', '
     for select in self.clause:
       dep = select.construct_node(self.subject, self.variants)
       if dep is None:
-        return Throw('Dependency {} is not satisfiable.'.format(select))
+        return Noop('Dependency {} is not satisfiable.'.format(select))
       dependencies.append(dep)
 
     # If all dependency Nodes are Return, execute the Node.
-    for dep_key in dependencies:
+    for dep_select, dep_key in zip(self.clause, dependencies):
       dep_state = dependency_states.get(dep_key, None)
       if dep_state is None or type(dep_state) == Waiting:
         return Waiting(dependencies)
       elif type(dep_state) == Return:
         dep_values.append(dep_state.value)
+      elif type(dep_state) == Noop:
+        if dep_select.optional:
+          dep_values.append(None)
+        else:
+          return Noop('Was missing (at least) input {}.'.format(dep_key))
       elif type(dep_state) == Throw:
-        return Throw('Dependency {} failed.'.format(dep_key))
+        return dep_state
       else:
         State.raise_unrecognized(dep_state)
-    return Return(self.func(*dep_values))
+    try:
+      return Return(self.func(*dep_values))
+    except Exception as e:
+      return Throw(e)
 
 
 class ProductGraph(object):
@@ -364,7 +391,7 @@ class ProductGraph(object):
     # A dict from Node to its computed value: if a Node hasn't been computed yet, it will not
     # be present here.
     self._node_results = dict()
-    # A dict from Node to list of dependency Nodes.
+    # Dicts from Nodes to sets of dependency/dependent Nodes.
     self._dependencies = defaultdict(set)
     self._dependents = defaultdict(set)
 
@@ -372,7 +399,7 @@ class ProductGraph(object):
     existing_state = self._node_results.get(node, None)
     if existing_state is not None:
       raise ValueError('Node {} is already completed:\n  {}\n  {}'.format(node, existing_state, state))
-    elif type(state) not in [Return, Throw]:
+    elif type(state) not in [Return, Throw, Noop]:
       raise ValueError('Cannot complete Node {} with state {}'.format(node, state))
     self._node_results[node] = state
 
@@ -389,7 +416,7 @@ class ProductGraph(object):
 
   def update_state(self, node, state):
     """Updates the Node with the given State."""
-    if type(state) in [Return, Throw]:
+    if type(state) in [Return, Throw, Noop]:
       self._set_state(node, state)
       return state
     elif type(state) == Waiting:
@@ -429,7 +456,7 @@ class ProductGraph(object):
     """Adds dependency edges from the given src Node to the given dependency Nodes.
 
     Executes cycle detection: if adding one of the given dependencies would create
-    a cycle, then the _source_ Node is marked as a Throw with an error indicating the
+    a cycle, then the _source_ Node is marked as a Noop with an error indicating the
     cycle path, and the dependencies are not introduced.
     """
     self.validate_node(node)
@@ -445,7 +472,7 @@ class ProductGraph(object):
       if cycle_path:
         # If a cycle is detected, don't introduce the dependencies, and instead fail the node.
         entries = ' ->\n  '.join(str(p) for p in cycle_path)
-        self._set_state(node, Throw('Cycle detected in path:\n  {} !!'.format(entries)))
+        self._set_state(node, Noop('Cycle detected in path:\n  {} !!'.format(entries)))
         return
 
     # Finally, add all deps.
@@ -454,6 +481,9 @@ class ProductGraph(object):
       self._dependents[dependency].add(node)
       # 'touch' the dependencies dict for this dependency, to ensure that an entry exists.
       self._dependencies[dependency]
+
+  def dependents(self):
+    return self._dependents
 
   def dependencies(self):
     return self._dependencies
@@ -467,7 +497,8 @@ class ProductGraph(object):
   def walk(self, roots, predicate=None):
     def _default_walk_predicate(entry):
       node, state = entry
-      return type(state) != Throw
+      cls = type(state)
+      return cls is not Noop
     predicate = predicate or _default_walk_predicate
 
     def _filtered_entries(nodes):
@@ -489,8 +520,8 @@ class ProductGraph(object):
         for e in _walk(dependencies):
           yield e
 
-    for node in _walk(_filtered_entries(roots)):
-      yield node
+    for entry in _walk(_filtered_entries(roots)):
+      yield entry
 
 
 class NodeBuilder(object):
@@ -500,19 +531,29 @@ class NodeBuilder(object):
   """
 
   @classmethod
-  def create(cls, tasks):
+  def create(cls, tasks, symbol_table_cls):
     """Indexes tasks by their output type."""
     serializable_tasks = defaultdict(set)
     for output_type, input_selects, task in tasks:
       serializable_tasks[output_type].add((task, tuple(input_selects)))
-    return cls(serializable_tasks)
+    literal_products = set(symbol_table_cls.table().values())
+    return cls(serializable_tasks, literal_products)
 
-  def __init__(self, tasks):
+  def __init__(self, tasks, literal_products):
     self._tasks = tasks
+    self._literal_products = literal_products
 
   def task_nodes(self, subject, product, variants):
+    # Tasks.
     for task, anded_clause in self._tasks[product]:
       yield TaskNode(subject, product, variants, task, anded_clause)
+    # An Address that might be resolved as a literal value from a build file.
+    # TODO: This defines a special case for Addresses by recognizing that they might be-a literal
+    # Product after resolution, and so it begins by attempting to resolve a Struct for
+    # a subject Address. This type of cast/conversion should likely be reified.
+    if isinstance(subject, Address) and product in self._literal_products:
+      struct_address = StructAddress(subject.spec_path, subject.target_name)
+      yield SelectNode(struct_address, Struct, None, None)
 
 
 class BuildRequest(object):
@@ -609,21 +650,110 @@ class Step(object):
     return 'Step({}, {})'.format(self._step_id, self._node)
 
 
-class LocalScheduler(object):
-  """A scheduler that expands a ProductGraph by executing user defined tasks.
+class GraphValidator(object):
+  """A concrete object that implements validation of a completed product graph.
 
-  # TODO(John Sirois): Allow for subject-less (target-less) goals.  Examples are clean-all,
-  # ng-killall, and buildgen.go.
+  TODO: If this abstraction seems useful, we should extract an interface and allow plugin
+  implementers to install their own. But currently the API isn't great: in particular, it
+  would be very, very helpful to be able to run validation _during_ graph execution as
+  subgraphs are completing. This would limit their performance impact, and allow for better
+  locality of errors.
   """
 
-  def __init__(self, products_by_goal, tasks):
+  def __init__(self, symbol_table_cls):
+    self._literal_types = dict()
+    for name, cls in symbol_table_cls.table().items():
+      self._literal_types[cls] = name
+
+  def _collect_consumed_inputs(self, product_graph, root):
+    """Walks successful nodes under the root for its subject, and returns all products used."""
+    consumed_inputs = set()
+    # Walk into successful nodes for the same subject under this root.
+    def predicate(entry):
+      node, state = entry
+      return root.subject == node.subject and type(state) is Return
+    # If a product was successfully selected, record it.
+    for ((node, _), _) in product_graph.walk([root], predicate=predicate):
+      if type(node) is SelectNode:
+        consumed_inputs.add(node.product)
+    return consumed_inputs
+
+  def _collect_partially_consumed_inputs(self, product_graph, consumed_inputs, root):
+    """Walks below a failed node and collects cases where additional literal products could be used.
+
+    Returns:
+      dict(subject, dict(tuple(input_product, output_product), list(tuple(task, missing_products))))
     """
-    :param products_by_goal: The products that are required for each goal name.
+    partials = defaultdict(lambda: defaultdict(list))
+    # Walk all nodes for the same subject under this root.
+    def predicate(entry):
+      node, state = entry
+      return root.subject == node.subject
+    for ((node, state), dependencies) in product_graph.walk([root], predicate=predicate):
+      # Look for unsatisfied TaskNodes with at least one unsatisfied dependency.
+      if type(node) is not TaskNode:
+        continue
+      if type(state) is not Noop:
+        continue
+      missing_products = {dep.product for dep, state in dependencies if type(state) == Noop}
+      if not missing_products:
+        continue
+
+      # If all unattainable products could have been specified as literal...
+      if any(product not in self._literal_types for product in missing_products):
+        continue
+
+      # There was at least one dep successfully (recursively) satisfied via a literal.
+      # TODO: this does multiple walks.
+      used_literal_deps = set()
+      for dep, _ in dependencies:
+        for product in self._collect_consumed_inputs(product_graph, dep):
+          if product in self._literal_types:
+            used_literal_deps.add(product)
+      if not used_literal_deps:
+        continue
+
+      # The partially consumed products were not fully consumed elsewhere.
+      if not (used_literal_deps - consumed_inputs):
+        continue
+
+      # Found a partially consumed input.
+      for used_literal_dep in used_literal_deps:
+        partials[node.subject][(used_literal_dep, node.product)].append((node.func, missing_products))
+    return partials
+
+  def validate(self, product_graph):
+    """Finds 'subject roots' in the product graph and invokes validation on each of them."""
+
+    # Locate roots: those who do not have any dependents for the same subject.
+    roots = set()
+    for node, dependents in product_graph.dependents().items():
+      if any(d.subject == node.subject for d in dependents):
+        # Node had a dependent for its subject: was not a root.
+        continue
+      roots.add(node)
+
+    # Raise if there were any partially consumed inputs.
+    for root in roots:
+      consumed = self._collect_consumed_inputs(product_graph, root)
+      partials = self._collect_partially_consumed_inputs(product_graph, consumed, root)
+      if partials:
+        raise PartiallyConsumedInputsError.create(self._literal_types, partials)
+
+
+class LocalScheduler(object):
+  """A scheduler that expands a ProductGraph by executing user defined tasks."""
+
+  def __init__(self, goals, symbol_table_cls, tasks):
+    """
+    :param goals: A dict from a goal name to a product type. A goal is just an alias for a
+           particular (possibly synthetic) product.
     :param tasks: A set of (output, input selection clause, task function) triples which
            is used to compute values in the product graph.
     """
-    self._products_by_goal = products_by_goal
-    self._node_builder = NodeBuilder.create(tasks)
+    self._goals = goals
+    self._graph_validator = GraphValidator(symbol_table_cls)
+    self._node_builder = NodeBuilder.create(tasks, symbol_table_cls)
     self._product_graph = ProductGraph()
     self._roots = set()
     self._step_id = -1
@@ -649,8 +779,8 @@ class LocalScheduler(object):
     # Determine the root products and subjects based on the request.
     root_subjects = [parse_variants(a) for a in build_request.addressable_roots]
     root_products = OrderedSet()
-    for goal in build_request.goals:
-      root_products.update(self._products_by_goal[goal])
+    for goal_name in build_request.goals:
+      root_products.add(self._goals[goal_name])
 
     # Roots are products that might be possible to produce for these subjects.
     return [SelectNode(s, p, v, None) for s, v in root_subjects for p in root_products]
@@ -666,10 +796,10 @@ class LocalScheduler(object):
     a tuple of (node_entry, dependency_node_entries).
 
     The given predicate is applied to entries, and eliminates the subgraphs represented by nodes
-    that don't match it. The default predicate eliminates all `Throw` subgraphs.
+    that don't match it. The default predicate eliminates all `Noop` subgraphs.
     """
-    for node in self._product_graph.walk(self._roots, predicate=predicate):
-      yield node
+    for entry in self._product_graph.walk(self._roots, predicate=predicate):
+      yield entry
 
   def root_entries(self):
     """Returns the roots for this scheduler as a dict from Node to State."""
@@ -740,8 +870,12 @@ class LocalScheduler(object):
             candidates.add(step.node)
 
     print('created {} total nodes in {} scheduling iterations and {} steps, '
-          'with {} nodes in the successful path.'.format(
+          'with {} nodes in the executed path.'.format(
             len(pg.dependencies()),
             scheduling_iterations,
             self._step_id,
             sum(1 for _ in pg.walk(self._roots))))
+
+  def validate(self):
+    """Validates the generated product graph with the configured GraphValidator."""
+    self._graph_validator.validate(self._product_graph)